+++ src/identity_set/drain/mod.rs
@@ -34,19 +34,18 @@ use allocator_api2::alloc::{Allocator, Global};
 use core::iter::FusedIterator;
 use core::ptr;
 
-/// Owning identity set iterator.
+/// Identity set drain.
 #[must_use]
 #[repr(transparent)]
-#[derive(Clone)]
-pub struct IntoIter<T, A: Allocator = Global> {
-	iter: identity_map::IntoIter<T, (), A>,
+pub struct Drain<'a, T, A: Allocator = Global> {
+	iter: identity_map::Drain<'a, T, (), A>,
 }
 
-impl<T, A: Allocator> IntoIter<T, A> {
-	/// Constructs a new, owning identity set iterator.
+impl<'a, T, A: Allocator> Drain<'a, T, A> {
+	/// Constructs a new identity set drain.
 	#[inline(always)]
-	pub(crate) fn new(set: IdentitySet<T, A>) -> Self {
-		let iter = set.into_map().into_iter();
+	pub(crate) fn new(set: &'a mut IdentitySet<T, A>) -> Self {
+		let iter = set.as_mut_map().drain();
 		Self { iter }
 	}
 
@@ -60,7 +59,7 @@ impl<T, A: Allocator> IntoIter<T, A> {
 	}
 }
 
-impl<T, A> Debug for IntoIter<T, A>
+impl<T, A> Debug for Drain<'_, T, A>
 where
 	T: Debug,
 	A: Allocator,
@@ -68,21 +67,13 @@ where
 	#[inline(always)]
 	fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
 		f
-			.debug_tuple("IntoIter")
+			.debug_tuple("Drain")
 			.field(&self.as_slice())
 			.finish()
 	}
 }
 
-impl<T, A: Allocator + Default> Default for IntoIter<T, A> {
-	#[inline(always)]
-	fn default() -> Self {
-		let iter = Default::default();
-		Self { iter }
-	}
-}
-
-impl<T, A: Allocator> Iterator for IntoIter<T, A> {
+impl<T, A: Allocator> Iterator for Drain<'_, T, A> {
 	type Item = T;
 
 	#[inline(always)]
@@ -96,13 +87,13 @@ impl<T, A: Allocator> Iterator for IntoIter<T, A> {
 	}
 }
 
-impl<T, A: Allocator> DoubleEndedIterator for IntoIter<T, A> {
+impl<T, A: Allocator> DoubleEndedIterator for Drain<'_, T, A> {
 	#[inline(always)]
 	fn next_back(&mut self) -> Option<Self::Item> {
 		self.iter.next().map(|(k, _)| k)
 	}
 }
 
-impl<T, A: Allocator> ExactSizeIterator for IntoIter<T, A> { }
+impl<T, A: Allocator> ExactSizeIterator for Drain<'_, T, A> { }
 
-impl<T, A: Allocator> FusedIterator for IntoIter<T, A> { }
+impl<T, A: Allocator> FusedIterator for Drain<'_, T, A> { }